chore(test): improves tests
