+++ lib/utils.js
@@ -45,12 +45,7 @@ const isInvalid = tab => tab.hidden || tab.disabled,
 			return tabs.find(isValid);
 		}
 		const selection = tabs.find(tab => tab.getAttribute('name') === selected);
-		if (selection == null) {
-			return tabs.find(isValid);
-		}
-		if (isValid(selection)) {
-			return selection;
-		}
+		return selection && isValid(selection) ? selection : tabs.find(isValid);
 	};
