+++ test/cosmoz-tabs-sizing.test.js
@@ -1,33 +1,27 @@
 import {
-	assert, html, fixture
+	assert, html, fixture, nextFrame
 } from '@open-wc/testing';
-import {
-	assert as sinonAssert,
-	spy
-} from 'sinon';
-import { flush } from '@polymer/polymer/lib/utils/flush';
 import '@polymer/iron-list';
 import '../cosmoz-tabs.js';
 
-suite('sizing', () => {
+suite('cosmoz-tabs sizing', () => {
 	suiteSetup(() => {
-		const documentContainer = document.createElement('template');
+		const customStyle = document.createElement('template');
+		customStyle.innerHTML = `<custom-style id="sizing-css"><style is="custom-style">
+			.item {
+				color: white;
+				overflow: hidden;
+			}
 
-		documentContainer.innerHTML = `<custom-style id="sizing-css"><style is="custom-style">
-		.item {
-			color: white;
-			overflow: hidden;
-		}
+			.item:nth-child(odd) {
+				background-color: green;
+			}
 
-		.item:nth-child(odd) {
-			background-color: green;
-		}
-
-		.item:nth-child(even) {
-			background-color: red;
-		}
-	</style></custom-style>`;
-		document.body.appendChild(documentContainer.content);
+			.item:nth-child(even) {
+				background-color: red;
+			}
+		</style></custom-style>`;
+		document.body.appendChild(customStyle.content);
 	});
 
 	suiteTeardown(() =>
@@ -41,7 +35,7 @@ suite('sizing', () => {
 					<iron-list>
 						<template>
 							<div class="item">
-								<div style="height: 100px;">[[item]]</div>
+								<div style="height: 100px;">[[ item ]]</div>
 							</div>
 						</template>
 					</iron-list>
@@ -57,22 +51,13 @@ suite('sizing', () => {
 				</cosmoz-tab>
 			</cosmoz-tabs>
 		`),
-			list = tabs.querySelector('iron-list'),
-			onResize = spy();
-
-		list.addEventListener('iron-resize', onResize);
-		flush();
+			list = tabs.querySelector('iron-list');
 
 		list.items = Array.from(Array(100).keys());
-		flush();
-		tabs._debouncer.flush();
-
-		list.removeEventListener('iron-resize', onResize);
+		await nextFrame();
 
 		assert.closeTo(list.getBoundingClientRect().height, 10000, 0.1);
 		assert.lengthOf(list.queryAllEffectiveChildren('.item'), 100);
-
-		sinonAssert.calledOnce(onResize);
 	});
 
 	test('explicitly sized tabs with element selected and flex list renders only a few items', async () => {
@@ -97,23 +82,14 @@ suite('sizing', () => {
 					</p>
 				</cosmoz-tab>
 			</cosmoz-tabs>`),
-			list = tabs.querySelector('iron-list'),
-			onResize = spy();
-
-		list.addEventListener('iron-resize', onResize);
-		flush();
+			list = tabs.querySelector('iron-list');
 
 		list.items = Array.from(Array(500).keys());
-		flush();
-		tabs._debouncer.flush();
-
-		list.removeEventListener('iron-resize', onResize);
+		await nextFrame();
 
 		assert.closeTo(list.getBoundingClientRect().height, 349, 0.1);
 		assert.isAbove(list.queryAllEffectiveChildren('.item').length, 3);
 		assert.isBelow(list.queryAllEffectiveChildren('.item').length, 12);
-
-		sinonAssert.calledOnce(onResize);
 	});
 
 	test('explicitly sized tabs and flex list updates size and renders only a few items upon selection', async () => {
@@ -139,24 +115,17 @@ suite('sizing', () => {
 					</cosmoz-tab>
 				</cosmoz-tabs>
 			`),
-			list = tabs.querySelector('iron-list'),
-			onResize = spy();
-
-		flush();
+			list = tabs.querySelector('iron-list');
 
 		assert.isUndefined(tabs.selected);
 		list.items = Array.from(Array(500).keys());
-		list.addEventListener('iron-resize', onResize);
 
 		tabs.selected = 'tab0';
-		flush();
-		tabs._debouncer.flush();
-		list.removeEventListener('iron-resize', onResize);
+		await nextFrame();
 
 		assert.closeTo(list.getBoundingClientRect().height, 349, 0.1);
 		assert.isAbove(list.queryAllEffectiveChildren('.item').length, 3);
 		assert.isBelow(list.queryAllEffectiveChildren('.item').length, 12);
 
-		sinonAssert.calledOnce(onResize);
 	});
 });
